move testnetbar
