+++ src/entries/popup/components/TestnetMode/TestnetBar/TestnetBarBackground.tsx
